New EJ1K (FlashGT) firmware 161114N4, Build chgs to support the Debian build

+++ src/kv/test/fvt_kv_utils.h
@@ -41,7 +41,7 @@
  *   load seed generated key/values into the ark using ark_set
  ******************************************************************************/
 void fvt_kv_utils_sload(ARK     *ark,
-                        uint32_t seed,
+                        uint64_t seed,
                         uint32_t klen,
                         uint32_t vlen,
                         uint32_t LEN);
@@ -51,7 +51,7 @@ void fvt_kv_utils_sload(ARK     *ark,
  *   query seed generated key/values in the ark using ark_get/ark_exists
  ******************************************************************************/
 void fvt_kv_utils_squery(ARK     *ark,
-                         uint32_t seed,
+                         uint64_t seed,
                          uint32_t klen,
                          uint32_t vlen,
                          uint32_t LEN);
@@ -61,7 +61,7 @@ void fvt_kv_utils_squery(ARK     *ark,
  *   query seed generated key/values in the ark using ark_get/ark_exists
  ******************************************************************************/
 void fvt_kv_utils_squery_empty(ARK     *ark,
-                               uint32_t seed,
+                               uint64_t seed,
                                uint32_t klen,
                                uint32_t LEN);
 
@@ -71,7 +71,7 @@ void fvt_kv_utils_squery_empty(ARK     *ark,
  *   delete seed generated key/values from the ark using ark_del
  ******************************************************************************/
 void fvt_kv_utils_sdel(ARK     *ark,
-                       uint32_t seed,
+                       uint64_t seed,
                        uint32_t klen,
                        uint32_t vlen,
                        uint32_t LEN);
@@ -82,7 +82,7 @@ void fvt_kv_utils_sdel(ARK     *ark,
  *   create db, set, get, exists, del, destroy seed generated key/values
  ******************************************************************************/
 void fvt_kv_utils_SGD_SLOOP(ARK     *ark,
-                           uint32_t seed,
+                           uint64_t seed,
                            uint32_t klen,
                            uint32_t vlen,
                            uint32_t len,
